# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Authentication Service API",
      "description": "API for managing authentication and user management in the system.",
      "version": "1.0"
    },
    "servers": [
      {
        "url": "http://localhost:8088",
        "description": "Local Development Server"
      },
      {
        "url": "https://authentication-service-193893744076.us-central1.run.app",
        "description": "Production Server"
      }
    ],
    "paths": {
      "/authentication-service/login": {
        "post": {
          "summary": "Login to the system",
          "tags": ["Authentication"],
          "operationId": "login",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  },
                  "required": ["username", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Login successful",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "token": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid credentials"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/authentication-service/users": {
        "get": {
          "summary": "Get all users (for admin only)",
          "tags": ["Authentication"],
          "operationId": "getAllUsers",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "List of all users",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            },
            "403": {
              "description": "Forbidden, only admin can access"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      }
    },
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT",
          "description": "JWT Authorization header using the Bearer scheme"
        }
      },
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "userId": {
              "type": "number"
            },
            "username": {
              "type": "string"
            },
            "password": {
              "type": "string"
            },
            "role": {
              "type": "string",
              "enum": ["admin", "operator"]
            },
            "email": {
              "type": "string"
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            }
          },
          "required": ["username", "password", "role"]
        }
      }
    }
  }
  
  